requirements.txt
customtkinter
CTkMessagebox
pytablericons
Pillow
requests
transformers
sentence-transformers
tk
